+++ src/layout/index.js
@@ -1,6 +1,6 @@
+export * from "./colorLegendLayout.js";
 export { default as dimensionsHeatTable } from "./heatTableDimensions.js";
 export { default as dimensionsHorizontalBarChart } from "./horizontalBarChartDimensions.js";
-export * from "./legendColorOrdinalLayout.js";
 export { default as layoutPopulationPyramid } from "./populationPyramidLayout.js";
 export { computeLayout as sankeyLayout, prepareData as sankeyPrepareData } from "./sankey.js";
 export { default as layoutSmallMultiples } from "./smallMultiples.js";